refactor(mirror): consolidate MetanormaToMirror into Transformer
MetanormaToMirror was the actual forward-conversion class — Transformer
was a 22-LOC facade that delegated to it. The facade naming was the
right public concept ("the transformer"), but the awkward
"MetanormaToMirror" name leaked an implementation-level concern.

Merge the body into Transformer so there is one forward-conversion
class with the right name. Transformer now serves both roles:
- Public entry point (#from_metanorma, #rewrite)
- Handler context object (#extract_blocks, #text_node, etc.)

Behavior unchanged. All 312 mirror specs pass.
